Added status files for each step in the -as-root script. Added GSConnect extension and gufw instructions to make it work.

+++ bookworm-setup.sh
@@ -76,7 +76,7 @@ if [ "$(id -u)" -eq 0 ]; then
 fi
 
 
-while getopts ':hife' opt; do
+while getopts ':hif' opt; do
 	case $opt in
 	h)
 		echo
@@ -102,12 +102,6 @@ while getopts ':hife' opt; do
 		echo
 		force=1
 		;;
-	e)
-		echo
-		echo 'Skip to Gnome extensions!'
-		echo
-		skip_to_ext=1
-		;;
 	\?)
 		echo
 		echo 'You entered an incorrect option...'
@@ -131,211 +125,225 @@ fi
 
 
 if [ ! -f "$status_dir/deb_setup_part_1" ] && [ ! -f "$status_dir/deb_setup_part_2" ]; then
-	if [ -z "$skip_to_ext" ]; then
-		if [ ! -f "$status_dir/reqs_confirmed" ]; then
-			echo 'This script automates some common settings that I use for'
-			echo 'every Debian installation while still allowing for some changes'
-			echo 'through interactive questions.  You will be asked to enter your'
-			echo 'password to sudo.'
-			echo
-			echo 'The script may require a few reboots, you will be prompted each'
-			echo 'time.  After the script reboots your system, please re-run the'
-			echo 'same script again and it should resume automatically.'
-
-
-			# Query user for requirements before proceeding
-			echo
-			echo 'Requirements:'
-			echo "    - Debian ${release_name^} installed, / and /home partitions set up as btrfs"
-			echo '    - Have patched fonts saved and unzipped in ~/fonts directory (default: Hack)'
-			echo '    - Have a stable Internet connection to download packages'
-			echo "    - Copied the files \"$release_name-config\", \"$release_name-gsettings.txt\", \"$release_name-dconf.txt\""
-			echo '          to a ~/Setup directory and customized them for this specific computer'
-			echo
-			read -p 'Have all of the above been completed? [y/N] '
-			if [ "${REPLY,}" != 'y' ]; then
-				echo
-				echo 'Please do those first, then run this script again!'
-				echo
-				exit
-			fi
-
-			# Create status directory
-			if [ ! -d "$status_dir" ]; then
-				echo
-				echo "Create status directory to hold script's state between reboots..."
-				confirm_cmd "mkdir $status_dir"
-			fi
-
-			touch "$status_dir/reqs_confirmed"
-			echo
-		fi
-
-
-		# Old versions of Wayland would not work well with screen-sharing for video conferencing, so I always prompted to switch to "Gnome on Xorg", however, it appears Wayland is doing better now (at least on Ubuntu), so this section is deprecated.
-		# If NOT installing NVIDIA driver, check for Wayland (nvidia-driver package will disable Wayland, at least for now)
-		# if [ -z "$(echo "${apt_installs[@]}" | grep nvidia)" ] && [ -n "$WAYLAND_DISPLAY" ]; then
-		# 	wayland=1
-		# fi
-
-
-		# Create temporary downloads directory
-		if [ ! -d "$script_dir/downloads" ]; then
-			echo
-			echo 'Create temporary downloads directory to hold packages...'
-			confirm_cmd "mkdir $script_dir/downloads"
-		fi
-
-
-		# Remove old configuration in .dotfiles
-		if [ -n "$rm_dotfiles" ] && [ ! -f "$status_dir/dotfiles_removed" ]; then
-			echo
-			echo 'Removing old .dotfiles (from prior Linux installation)...'
-			confirm_cmd "sudo rm -rf $HOME/.*"
-			confirm_cmd "cp -av /etc/skel/. $HOME/"
-			touch "$status_dir/dotfiles_removed"
-			echo
-		fi
-
-
-		# Disable suspend while on AC power
-		if [ ! -f "$status_dir/disabled_ac_suspend" ]; then
-			echo
-			echo 'Disable suspend while on AC power...'
-			confirm_cmd "gsettings set org.gnome.settings-daemon.plugins.power sleep-inactive-ac-type 'nothing'"
-			touch "$status_dir/disabled_ac_suspend"
-		fi
-
-
-		# Run commands as root (with sudo)
-		# sudo home="$HOME" interactive="$interactive" wayland="$wayland" bash "$release_name"-as-root
-		sudo home="$HOME" interactive="$interactive" bash "$release_name"-as-root
-
-
-		# Change default apps (VLC instead of Videos/Totem)
-		# I can't seem to find where these default applications are saved to change them programmatically,
-		# so I'm resorting to the GUI...  If anybody knows where I can change this in gsettings or dconf, let me know!
+	if [ ! -f "$status_dir/reqs_confirmed" ]; then
+		echo 'This script automates some common settings that I use for'
+		echo 'every Debian installation while still allowing for some changes'
+		echo 'through interactive questions.  You will be asked to enter your'
+		echo 'password to sudo.'
 		echo
-		echo 'The default Gnome Videos player never seems to do very well.  Switch the default'
-		echo 'video player to VLC, then close the window to continue the setup script...'
+		echo 'The script may require a few reboots, you will be prompted each'
+		echo 'time.  After the script reboots your system, please re-run the'
+		echo 'same script again and it should resume automatically.'
+
+
+		# Query user for requirements before proceeding
 		echo
-		read -p 'Press ENTER to open Display Settings...'
-		confirm_cmd 'gnome-control-center default-apps'
+		echo 'Requirements:'
+		echo "    - Debian ${release_name^} installed, / and /home partitions set up as btrfs"
+		echo '    - Have patched fonts saved and unzipped in ~/fonts directory (default: Hack)'
+		echo '    - Have a stable Internet connection to download packages'
+		echo "    - Copied the files \"$release_name-config\", \"$release_name-gsettings.txt\", \"$release_name-dconf.txt\""
+		echo '          to a ~/Setup directory and customized them for this specific computer'
+		echo
+		read -p 'Have all of the above been completed? [y/N] '
+		if [ "${REPLY,}" != 'y' ]; then
+			echo
+			echo 'Please do those first, then run this script again!'
+			echo
+			exit
+		fi
+		touch "$status_dir/reqs_confirmed"
+		echo
 	fi
+
+
+	# Old versions of Wayland would not work well with screen-sharing for video conferencing, so I always prompted to switch to "Gnome on Xorg", however, it appears Wayland is doing better now (at least on Ubuntu), so this section is deprecated.
+	# If NOT installing NVIDIA driver, check for Wayland (nvidia-driver package will disable Wayland, at least for now)
+	# if [ -z "$(echo "${apt_installs[@]}" | grep nvidia)" ] && [ -n "$WAYLAND_DISPLAY" ]; then
+	# 	wayland=1
+	# fi
+
+
+	# Create status directory
+	if [ ! -d "$status_dir" ]; then
+		echo
+		echo "Create status directory to hold script's state between reboots..."
+		confirm_cmd "mkdir $status_dir"
+	fi
+
+
+	# Create temporary downloads directory
+	if [ ! -d "$script_dir/downloads" ]; then
+		echo
+		echo 'Create temporary downloads directory to hold packages...'
+		confirm_cmd "mkdir $script_dir/downloads"
+	fi
+
+
+	# Remove old configuration in .dotfiles
+	if [ -n "$rm_dotfiles" ] && [ ! -f "$status_dir/dotfiles_removed" ]; then
+		echo
+		echo 'Removing old .dotfiles (from prior Linux installation)...'
+		confirm_cmd "sudo rm -rf $HOME/.*"
+		confirm_cmd "cp -av /etc/skel/. $HOME/"
+		touch "$status_dir/dotfiles_removed"
+		echo
+	fi
+
+
+	# Inhibit suspend while on AC power
+	if [ ! -f "$status_dir/inhibit_ac_suspend" ]; then
+		echo
+		echo 'Disable suspend while on AC power...'
+		confirm_cmd "gsettings set org.gnome.settings-daemon.plugins.power sleep-inactive-ac-type 'nothing'"
+		touch "$status_dir/inhibit_ac_suspend"
+	fi
+
+
+	# Run commands as root (with sudo)
+	# sudo home="$HOME" interactive="$interactive" wayland="$wayland" bash "$release_name"-as-root
+	sudo home="$HOME" interactive="$interactive" bash "$release_name"-as-root
+
+
+	# Change default apps (VLC instead of Videos/Totem)
+	# I can't seem to find where these default applications are saved to change them programmatically,
+	# so I'm resorting to the GUI...  If anybody knows where I can change this in gsettings or dconf, let me know!
+	echo
+	echo 'The default Gnome Videos player never seems to do very well.  Switch the default'
+	echo 'video player to VLC, then close the window to continue the setup script...'
+	echo
+	read -p 'Press ENTER to open Display Settings...'
+	confirm_cmd 'gnome-control-center default-apps'
 
 
 	# Install Gnome extensions
-	echo
-	echo 'Installing Gnome extensions...'
-	echo
-	gnome_ver=$(gnome-shell --version | cut -d' ' -f3)
-	base_url='https://extensions.gnome.org'
-	for extension in "${extension_urls[@]}"; do
-		ext_uuid=$(curl -s $extension | grep -oP 'data-uuid="\K[^"]+')
-		info_url="$base_url/extension-info/?uuid=$ext_uuid&shell_version=$gnome_ver"
-		download_url="$base_url$(curl -s "$info_url" | sed -e 's/.*"download_url": "\([^"]*\)".*/\1/')"
-		confirm_cmd "curl -L '$download_url' > '$script_dir/downloads/$ext_uuid.zip'"
-		ext_dir="$HOME/.local/share/gnome-shell/extensions/$ext_uuid"
-		confirm_cmd "gnome-extensions install $script_dir/downloads/$ext_uuid.zip"
-		# DEPRECATED # Move all indicators to the right of the system-monitor indicator on the panel
-		# if [ -z $(echo "$ext_uuid" | grep 'system-monitor') ]; then
-		# 	confirm_cmd "sed -i 's/\(Main.panel.addToStatusArea([^,]*,[^,]*\)\(, [0-9]\)\?);/\1, 2);/' $ext_dir/extension.js"
-		# fi
-	done
-	echo
+	if [ -n "${extension_urls[*]}" ] && [ ! -f "$status_dir/extensions_installed" ]; then
+		echo
+		echo 'Installing Gnome extensions...'
+		echo
+		gnome_ver=$(gnome-shell --version | cut -d' ' -f3)
+		base_url='https://extensions.gnome.org'
+		for extension in "${extension_urls[@]}"; do
+			ext_uuid=$(curl -s $extension | grep -oP 'data-uuid="\K[^"]+')
+			info_url="$base_url/extension-info/?uuid=$ext_uuid&shell_version=$gnome_ver"
+			download_url="$base_url$(curl -s "$info_url" | sed -e 's/.*"download_url": "\([^"]*\)".*/\1/')"
+			confirm_cmd "curl -L '$download_url' > '$script_dir/downloads/$ext_uuid.zip'"
+			ext_dir="$HOME/.local/share/gnome-shell/extensions/$ext_uuid"
+			confirm_cmd "gnome-extensions install $script_dir/downloads/$ext_uuid.zip"
+			# DEPRECATED # Move all indicators to the right of the system-monitor indicator on the panel
+			# if [ -z $(echo "$ext_uuid" | grep 'system-monitor') ]; then
+			# 	confirm_cmd "sed -i 's/\(Main.panel.addToStatusArea([^,]*,[^,]*\)\(, [0-9]\)\?);/\1, 2);/' $ext_dir/extension.js"
+			# fi
+		done
+		touch "$status_dir/extensions_installed"
+		echo
+	fi
 
 
 	# Set up fonts
-	echo
-	echo 'Setting up links to detect fonts...'
-	echo
-	confirm_cmd "ln -s $HOME/fonts $HOME/.local/share/fonts"
-	confirm_cmd 'fc-cache -fv'
-	echo
+	if [ ! -f "$status_dir/fonts_installed" ]; then
+		echo
+		echo 'Setting up links to detect fonts...'
+		echo
+		confirm_cmd "ln -s $HOME/fonts $HOME/.local/share/fonts"
+		confirm_cmd 'fc-cache -fv'
+		touch "$status_dir/fonts_installed"
+		echo
+	fi
 
 
 	# Load gsettings
-	echo
-	echo "Applying $release_name-gsettings.txt to Gnome..."
-	echo
-	# confirm_cmd in while loop won't work since it also uses 'read', so
-	# confirmation must be asked beforehand if $interactive is true
-	if [ -n "$interactive" ]; then
-		echo -e "Load all settings from gsettings.txt using:\n    $ gsettings set \$schema \$key \"\$val\""
-		read -p 'Proceed? [Y/n] '
+	if [ ! -f "$status_dir/gsettings_loaded" ]; then
+		echo
+		echo "Applying $release_name-gsettings.txt to Gnome..."
+		echo
+		# confirm_cmd in while loop won't work since it also uses 'read', so
+		# confirmation must be asked beforehand if $interactive is true
+		if [ -n "$interactive" ]; then
+			echo -e "Load all settings from gsettings.txt using:\n    $ gsettings set \$schema \$key \"\$val\""
+			read -p 'Proceed? [Y/n] '
+		fi
+		if [ -z "$interactive" ] || [ -z "$REPLY" ] || [ "${REPLY,}" == 'y' ]; then
+			while read -r schema key val; do
+				echo -e "\nExecuting command...\n    $ gsettings set $schema $key \"$val\"\n"
+				gsettings set $schema $key "$val"
+			done < "$gsettings_path"
+		fi
+		touch "$status_dir/gsettings_loaded"
+		echo
 	fi
-	if [ -z "$interactive" ] || [ -z "$REPLY" ] || [ "${REPLY,}" == 'y' ]; then
-		while read -r schema key val; do
-			echo -e "\nExecuting command...\n    $ gsettings set $schema $key \"$val\"\n"
-			gsettings set $schema $key "$val"
-		done < "$gsettings_path"
-	fi
-	echo
 
 
 	# Load dconf settings
-	echo
-	echo "Applying $release_name-dconf.txt to Gnome..."
-	echo
-	confirm_cmd "dconf load / < $dconf_settings_path"
-	echo
+	if [ ! -f "$status_dir/dconf_loaded" ]; then
+		echo
+		echo "Applying $release_name-dconf.txt to Gnome..."
+		echo
+		confirm_cmd "dconf load / < $dconf_settings_path"
+		touch "$status_dir/dconf_loaded"
+		echo
+	fi
 
 
 	# Ignore suspend on closing lid tweak
-	if [ -n "$ignore_lid_switch" ]; then
+	if [ -n "$ignore_lid_switch" ] && [ ! -f "$status_dir/lid_tweak_installed" ]; then
 		echo
 		echo 'Applying tweak to ignore suspend on lid closing...'
 		if [ ! -d "$HOME/.config/autostart" ]; then
 			confirm_cmd "mkdir $HOME/.config/autostart"
 		fi
 		confirm_cmd 'echo -e "[Desktop Entry]\\nType=Application\\nName=ignore-lid-switch-tweak\\nExec=/usr/libexec/gnome-tweak-tool-lid-inhibitor\\n" > $HOME/.config/autostart/ignore-lid-switch-tweak.desktop'
+		touch "$status_dir/lid_tweak_installed"
 		echo
 	fi
 
 
 	# Create startup application to move indicators
-	if [ -n "$move_indicators" ]; then
+	if [ -n "$move_system_monitor" ] && [ ! -f "$status_dir/move_system_monitor_installed" ]; then
 		echo
 		echo 'Creating startup application to move Gnome extension system-monitor to the left of all indicators...'
 		if [ ! -d "$HOME/.config/autostart" ]; then
 			confirm_cmd "mkdir $HOME/.config/autostart"
 		fi
 		confirm_cmd "echo -e \"[Desktop Entry]\\\\nType=Application\\\\nName=Move system-monitor indicator\\\\nComment=Moves user-installed Gnome extension system-monitor to the left of all indicators (updates periodically move it back to the middle)\\\\nExec=/usr/local/bin/move-system-monitor\\\\n\" > $HOME/.config/autostart/move-system-monitor.desktop"
+		touch "$status_dir/move_system_monitor_installed"
 		echo
 	fi
 
 
 	# Set up Neovim
-	# Clone neovim-config from GitHub
-	echo
-	echo 'Setting up Neovim config (init.vim)...'
-	echo
-	if [ ! -d "$HOME/dotfiles" ]; then
-		confirm_cmd "mkdir $HOME/dotfiles"
+	if [ -n "$(contains apt_installs neovim)" ] && [ ! -f "$status_dir/neovim_installed" ]; then
+		# Clone neovim-config from GitHub
+		echo
+		echo 'Setting up Neovim config (init.vim)...'
+		echo
+		if [ ! -d "$HOME/dotfiles" ]; then
+			confirm_cmd "mkdir $HOME/dotfiles"
+		fi
+		confirm_cmd "git -C $HOME/dotfiles/ clone https://github.com/tedlava/neovim-config.git"
+		confirm_cmd "mkdir $HOME/.config/nvim"
+		confirm_cmd "ln -s $HOME/dotfiles/neovim-config/init.vim $HOME/.config/nvim/"
+		echo
+		echo 'Installing vim-plug into Neovim...'
+		confirm_cmd "sh -c 'curl -fLo \"${XDG_DATA_HOME:-$HOME/.local/share}\"/nvim/site/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim'"
+		echo
+		echo 'About to install Neovim plugins.  When Neovim is finished, please exit'
+		echo 'Neovim by typing ":qa" and pressing ENTER.'
+		echo
+		echo '    *** Do NOT close the terminal window! ***'
+		echo
+		read -p 'Press ENTER to proceed with Neovim plugin installation. '
+		confirm_cmd "nvim -c PlugInstall"
+		echo
+
+		# Load Nautilus mime types for Neovim
+		echo
+		echo 'Loading Nautilus mime types (open all text files with Neovim)...'
+		echo
+		confirm_cmd "cp -av mimeapps.list $HOME/.config/"
+		touch "$status_dir/neovim_installed"
+		echo
 	fi
-	confirm_cmd "git -C $HOME/dotfiles/ clone https://github.com/tedlava/neovim-config.git"
-	confirm_cmd "mkdir $HOME/.config/nvim"
-	confirm_cmd "ln -s $HOME/dotfiles/neovim-config/init.vim $HOME/.config/nvim/"
-	echo
-	echo 'Installing vim-plug into Neovim...'
-	confirm_cmd "sh -c 'curl -fLo \"${XDG_DATA_HOME:-$HOME/.local/share}\"/nvim/site/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim'"
-	echo
-	echo 'About to install Neovim plugins.  When Neovim is finished, please exit'
-	echo 'Neovim by typing ":qa" and pressing ENTER.'
-	echo
-	echo '    *** Do NOT close the terminal window! ***'
-	echo
-	read -p 'Press ENTER to proceed with Neovim plugin installation. '
-	confirm_cmd "nvim -c PlugInstall"
-	echo
-
-
-	# Load Nautilus mime types for Neovim
-	echo
-	echo 'Loading Nautilus mime types (open all text files with Neovim)...'
-	echo
-	confirm_cmd "cp -av mimeapps.list $HOME/.config/"
-	echo
 
 
 	# Reboot
@@ -361,7 +369,7 @@ if [ ! -f "$status_dir/deb_setup_part_1" ] && [ ! -f "$status_dir/deb_setup_part
 
 elif [ -f "$status_dir/deb_setup_part_1" ] && [ ! -f "$status_dir/deb_setup_part_2" ]; then
 	# Enable Gnome extensions
-	if [ -n "${extension_urls[*]}" ]; then
+	if [ -n "${extension_urls[*]}" ] && [ ! -f "$status_dir/extensions_enabled" ]; then
 		echo
 		echo 'Enabling recently installed Gnome extensions...'
 		echo
@@ -369,24 +377,29 @@ elif [ -f "$status_dir/deb_setup_part_1" ] && [ ! -f "$status_dir/deb_setup_part
 			ext_uuid=$(curl -s $extension | grep -oP 'data-uuid="\K[^"]+')
 			confirm_cmd "gnome-extensions enable ${ext_uuid}"
 		done
+		touch "$status_dir/extensions_enabled"
 	fi
 
 
 	# Open display settings to possibly change scaling
-	echo
-	echo 'Newer computers with HiDPI displays may need to adjust scaling settings. When'
-	echo 'you are finished, close the window to continue with the setup script...'
-	echo
-	read -p 'Press ENTER to open Display Settings...'
-	confirm_cmd 'gnome-control-center display'
+	if [ ! -f "$status_dir/display_settings" ]; then
+		echo
+		echo 'Newer computers with HiDPI displays may need to adjust scaling settings. When'
+		echo 'you are finished, close the window to continue with the setup script...'
+		echo
+		read -p 'Press ENTER to open Display Settings...'
+		confirm_cmd 'gnome-control-center display'
+		touch "$status_dir/display_settings"
+	fi
 
 
 	# Install flatpaks
-	if [ -n "${flatpaks[*]}" ]; then
+	if [ -n "${flatpaks[*]}" ] && [ ! -f "$status_dir/flatpaks_installed" ]; then
 		echo
 		echo 'Installing Flatpak applications...'
 		echo
 		confirm_cmd "flatpak -y install ${flatpaks[@]}"
+		touch "$status_dir/flatpaks_installed"
 		echo
 	fi
 
@@ -407,10 +420,13 @@ elif [ -f "$status_dir/deb_setup_part_1" ] && [ ! -f "$status_dir/deb_setup_part
 
 
 	# Create timeshift snapshot after setup script is complete
-	echo
-	echo 'Create a timeshift snapshot in case you screw up this awesome setup...'
-	confirm_cmd "sudo timeshift --create --comments 'Debian ${release_name^} setup script completed' --yes"
-	echo
+	if [ ! -f "$status_dir/final_snapshot" ]; then
+		echo
+		echo 'Create a timeshift snapshot in case you screw up this awesome setup...'
+		confirm_cmd "sudo timeshift --create --comments 'Debian ${release_name^} setup script completed' --yes"
+		touch "$status_dir/final_snapshot"
+		echo
+	fi
 
 
 	# Settings to fix after this script...
